# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "summary": "PUBLIC: Register new user",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User successfully registered."
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "summary": "PUBLIC: Login user",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User successfully logged in."
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/refresh": {
      "post": {
        "operationId": "AuthController_refreshToken",
        "summary": "PUBLIC: Refresh access token",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Token successfully refreshed."
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "summary": "Logout user",
        "parameters": [],
        "responses": {
          "200": {
            "description": "User successfully logged out."
          }
        },
        "tags": [
          "auth"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/auth/me": {
      "get": {
        "operationId": "AuthController_getProfile",
        "summary": "Get current user profile",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Returns the current user profile."
          }
        },
        "tags": [
          "auth"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
...
    "/airport": {
      "post": {
        "operationId": "AirportController_create",
        "summary": "ADMIN: Create an airport",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAirportDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Create an airport operation completed successfully."
          }
        },
        "tags": [
          "airport"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      },
      "get": {
        "operationId": "AirportController_findAll",
        "summary": "PUBLIC: Get all airports",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Returns a list of all airports"
          }
        },
        "tags": [
          "airport"
        ]
      }
    },
    "/airport/{id}": {
      "get": {
        "operationId": "AirportController_findOne",
        "summary": "PUBLIC: Find an airport by id",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns a single airport by its ID"
          }
        },
        "tags": [
          "airport"
        ]
      },
      "patch": {
        "operationId": "AirportController_update",
        "summary": "ADMIN: Update an airport",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAirportDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Update an airport operation completed successfully."
          }
        },
        "tags": [
          "airport"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      },
      "delete": {
        "operationId": "AirportController_remove",
        "summary": "ADMIN: Delete an airport",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Delete an airport operation completed successfully."
          }
        },
        "tags": [
          "airport"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
...
    "/article": {
      "post": {
        "operationId": "ArticleController_create",
        "summary": "ADMIN: Create an article",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateArticleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Create an article operation completed successfully."
          }
        },
        "tags": [
          "article"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      },
      "get": {
        "operationId": "ArticleController_findAll",
        "summary": "ADMIN: Returns all articles",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Returns all articles operation completed successfully."
          }
        },
        "tags": [
          "article"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/article/published": {
      "get": {
        "operationId": "ArticleController_findPublished",
        "summary": "PUBLIC: Get all published articles",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Returns all published articles"
          },
          "400": {
            "description": "Bad request"
          },
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Article"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "article"
        ]
      }
    },
    "/article/{id}": {
      "get": {
        "operationId": "ArticleController_findOne",
        "summary": "PUBLIC: Get article by id",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns a single article"
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Article not found"
          },
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Article"
                }
              }
            }
          }
        },
        "tags": [
          "article"
        ]
      },
      "patch": {
        "operationId": "ArticleController_update",
        "summary": "ADMIN: Update article by id",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateArticleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Update article by id operation completed successfully."
          }
        },
        "tags": [
          "article"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      },
      "delete": {
        "operationId": "ArticleController_remove",
        "summary": "ADMIN: Delete article by id",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Delete article by id operation completed successfully."
          }
        },
        "tags": [
          "article"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    }
  },
  "info": {
    "title": "QAirline",
    "description": "The QAirline API description",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
    "schemas": {
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "fullname": {
            "type": "string",
            "description": "Full name of the user",
            "example": "Nguyễn Văn A",
            "minLength": 6,
            "maxLength": 50
          },
          "phone_number": {
            "type": "string",
            "description": "Vietnam phone number",
            "example": "0123456789",
            "minLength": 10,
            "maxLength": 10
          },
          "email": {
            "type": "string",
            "description": "Email address",
            "example": "example@gmail.com"
          },
          "password_hash": {
            "type": "string",
            "description": "Password with specific requirements",
            "example": "StrongP@ssw0rd!"
          },
          "role": {
            "type": "string",
            "description": "User role",
            "enum": [
              "ADMIN",
              "USER"
            ],
            "default": "USER"
          },
          "birthday": {
            "type": "string",
            "description": "User's date of birth",
            "example": "1990-01-01",
            "format": "date"
          },
          "gender": {
            "type": "string",
            "description": "Gender of the user",
            "enum": [
              "MALE",
              "FEMALE"
            ]
          }
        },
        "required": [
          "fullname",
          "phone_number",
          "email",
          "password_hash",
          "birthday",
          "gender"
        ]
      },
...
      "LoginDto": {
        "type": "object",
        "properties": {
          "phone_number": {
            "type": "string",
            "description": "phone number",
            "example": "0123456789"
          },
          "password": {
            "type": "string",
            "description": "password",
            "example": "StrongP@ssw0rd!"
          }
        },
        "required": [
          "phone_number",
          "password"
        ]
      },
      "RefreshTokenDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "description": "Refresh token",
            "example": "eyJhbGci"
          }
        },
        "required": [
          "refreshToken"
        ]
      },
...
      "CreateAirportDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "The code of the airport",
            "example": "JFK"
          },
          "name": {
            "type": "string",
            "description": "The name of the airport",
            "example": "John F. Kennedy International Airport"
          },
          "city": {
            "type": "string",
            "description": "The city where the airport is located",
            "example": "New York"
          },
          "country": {
            "type": "string",
            "description": "The country where the airport is located",
            "example": "United States"
          }
        },
        "required": [
          "code",
          "name",
          "city",
          "country"
        ]
      },
      "UpdateAirportDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "The code of the airport",
            "example": "JFK"
          },
          "name": {
            "type": "string",
            "description": "The name of the airport",
            "example": "John F. Kennedy International Airport"
          },
          "city": {
            "type": "string",
            "description": "The city where the airport is located",
            "example": "New York"
          },
          "country": {
            "type": "string",
            "description": "The country where the airport is located",
            "example": "United States"
          }
        }
      },
...
      "CreateArticleDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "The title of the article",
            "example": "How to create a NestJS application"
          },
          "description": {
            "type": "string",
            "description": "The description of the article",
            "example": "A step-by-step guide to create a NestJS application"
          },
          "content": {
            "type": "string",
            "description": "The content of the article",
            "example": "This is the content of the article"
          },
          "image_url": {
            "type": "string",
            "description": "The image URL of the article",
            "example": "https://example.com/image.jpg"
          }
        },
        "required": [
          "title",
          "description",
          "content",
          "image_url"
        ]
      },
      "Article": {
        "type": "object",
        "properties": {}
      },
      "UpdateArticleDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "The title of the article",
            "example": "How to create a NestJS application"
          },
          "description": {
            "type": "string",
            "description": "The description of the article",
            "example": "A step-by-step guide to create a NestJS application"
          },
          "content": {
            "type": "string",
            "description": "The content of the article",
            "example": "This is the content of the article"
          },
          "image_url": {
            "type": "string",
            "description": "The image URL of the article",
            "example": "https://example.com/image.jpg"
          },
          "status": {
            "type": "string",
            "description": "The status of the article",
            "example": "PUBLISHED"
          }
        },
        "required": [
          "title",
          "description",
          "content",
          "image_url",
          "status"
        ]
      }
    }
  }
}